requirements.txt
click==7.1.2
Flask==1.1.2
IMDbPY==2020.9.25
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
SQLAlchemy==1.3.20
Werkzeug==1.0.1
gunicorn==19.9.0
